+++ simconvoi.cc
@@ -218,7 +218,7 @@ DBG_MESSAGE("convoi_t::~convoi_t()", "destroying %d, %p", self.get_id(), this);
 	// force asynchronous recalculation
 	if(fpl) {
 		if(!fpl->ist_abgeschlossen()) {
-			destroy_win((long)fpl);
+			destroy_win((ptrdiff_t)fpl);
 		}
 		if (!fpl->empty() && !line.is_bound()) {
 			welt->set_schedule_counter();
@@ -587,7 +587,7 @@ void convoi_t::set_name(const char *name, bool with_new_id)
 		if(  ground  ) {
 			const depot_t *const depot = ground->get_depot();
 			if(  depot  ) {
-				depot_frame_t *const frame = dynamic_cast<depot_frame_t *>( win_get_magic( (long)depot ) );
+				depot_frame_t *const frame = dynamic_cast<depot_frame_t *>( win_get_magic( (ptrdiff_t)depot ) );
 				if(  frame  ) {
 					frame->reset_convoy_name( self );
 				}
@@ -1600,7 +1600,7 @@ bool convoi_t::set_schedule(schedule_t * f)
 		}
 		// destroy a possibly open schedule window
 		if(  fpl  &&  !fpl->ist_abgeschlossen()  ) {
-			destroy_win((long)fpl);
+			destroy_win((ptrdiff_t)fpl);
 			delete fpl;
 		}
 		fpl = f;
@@ -2324,7 +2324,7 @@ void convoi_t::zeige_info()
 				if(  depot  ) {
 					depot->zeige_info();
 					// try to activate this particular convoy in the depot
-					depot_frame_t *const frame = dynamic_cast<depot_frame_t *>( win_get_magic( (long)depot ) );
+					depot_frame_t *const frame = dynamic_cast<depot_frame_t *>( win_get_magic( (ptrdiff_t)depot ) );
 					if(  frame  ) {
 						frame->activate_convoi(self);
 					}
@@ -2468,7 +2468,7 @@ void convoi_t::open_schedule_window( bool show )
 
 	if(  show  ) {
 		// Fahrplandialog oeffnen
-		create_win( new fahrplan_gui_t(fpl,get_besitzer(),self), w_info, (long)fpl );
+		create_win( new fahrplan_gui_t(fpl,get_besitzer(),self), w_info, (ptrdiff_t)fpl );
 		// TODO: what happens if no client opens the window??
 	}
 	fpl->eingabe_beginnen();
@@ -2811,7 +2811,7 @@ void convoi_t::destroy()
 	state = SELF_DESTRUCT;
 
 	if(fpl!=NULL  &&  !fpl->ist_abgeschlossen()) {
-		destroy_win((long)fpl);
+		destroy_win((ptrdiff_t)fpl);
 	}
 
 	if(  line.is_bound()  ) {
